feat: folder structure for service layer pattern
