+++ .vscode/settings.json
@@ -1,0 +1,5 @@
+{
+    "gopls": {
+        "ui.semanticTokens": true
+    }
+}
