--- README.md
@@ -1,2 +1,0 @@
-# cong-manager
-Manage a congregation with ease.
