+++ backend/README.md
@@ -1,0 +1,10 @@
+# cong-manager
+
+Manage a congregation with ease.
+
+## Development
+
+### To start
+
+- `go run cmd/main.go`
+
