+++ backend/internal/services/test_service.go
@@ -1,0 +1,7 @@
+package services
+
+import "fmt"
+
+func Hello() {
+	fmt.Println("hi")
+}
